% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umweltapir-package.R
\docType{package}
\name{umweltapir-package}
\alias{umweltapir}
\alias{umweltapir-package}
\title{umweltapir: Access Umwelt.Info API}
\description{
Provides an R-based access to the datasets including their resources from the portal \url{https://umwelt.info}. The package allows for an easy integration of those datasets into your R-based workflows. The functionality of the package mirrors the web-based access as provided at \url{https://umwelt.info}. You can use the same queries and get the same datasets by accessing our API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.opencode.de/umwelt-info/packages/-/tree/main/umweltapir?ref_type=heads}
  \item Report bugs at \url{https://gitlab.opencode.de/umwelt-info/packages/-/work_items}
}

}
\author{
\strong{Maintainer}: Johannes Vogel \email{johannes.vogel@uba.de}

Authors:
\itemize{
  \item Maximilian Berthold \email{maximilian.berthold@uba.de}
}

Other contributors:
\itemize{
  \item Luise Quoß \email{luise.quoss@uba.de} [contributor]
  \item Nationales Zentrum für Umwelt- und Naturschutzinformationen \email{umwelt.info@uba.de} [copyright holder]
}

}
\keyword{internal}
